refactor(sequencer): move sequencer logic to dedicated crate
