+++ crates/sequencer/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "zone-sequencer"
+description = "Tempo Zone sequencer settlement and withdrawal tasks"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# tempo
+tempo-alloy.workspace = true
+tempo-zone-contracts = { workspace = true, features = ["std", "serde", "rpc"] }
+
+# reth
+reth-metrics.workspace = true
+
+# alloy
+alloy-consensus.workspace = true
+alloy-network.workspace = true
+alloy-primitives.workspace = true
+alloy-provider = { workspace = true, features = ["ws"] }
+alloy-rlp.workspace = true
+alloy-rpc-client.workspace = true
+alloy-signer-local.workspace = true
+alloy-sol-types.workspace = true
+alloy-transport.workspace = true
+
+# misc
+eyre.workspace = true
+futures.workspace = true
+metrics.workspace = true
+parking_lot.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+[dev-dependencies]
+alloy-rpc-types-eth.workspace = true
+serde_json.workspace = true
+tempo-primitives.workspace = true
+tokio-tungstenite.workspace = true
