+++ crates/tempo-zone/Cargo.toml
@@ -28,12 +28,12 @@ zone-payload.workspace = true
 zone-precompiles = { workspace = true, features = ["std"] }
 zone-primitives = { workspace = true, features = ["std", "serde"] }
 zone-rpc.workspace = true
+zone-sequencer.workspace = true
 
 # reth
 reth-chainspec.workspace = true
 reth-eth-wire-types.workspace = true
 reth-evm.workspace = true
-reth-metrics.workspace = true
 reth-node-api.workspace = true
 reth-node-builder.workspace = true
 reth-payload-builder.workspace = true
@@ -56,19 +56,16 @@ alloy-network.workspace = true
 alloy-primitives.workspace = true
 alloy-provider.workspace = true
 alloy-rpc-client.workspace = true
-alloy-rlp.workspace = true
 alloy-rpc-types-engine.workspace = true
 alloy-rpc-types-eth.workspace = true
 alloy-signer-local.workspace = true
 alloy-sol-types.workspace = true
-alloy-transport.workspace = true
 
 # revm
 revm.workspace = true
 
 # misc
 clap = { workspace = true, optional = true }
-metrics.workspace = true
 k256.workspace = true
 reth-consensus = { workspace = true, optional = true }
 reth-ethereum = { workspace = true, features = [
@@ -78,7 +75,6 @@ reth-ethereum = { workspace = true, features = [
 reth-tracing = { workspace = true, optional = true }
 eyre.workspace = true
 futures.workspace = true
-parking_lot.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 
@@ -95,6 +91,7 @@ alloy = { workspace = true, features = [
 alloy-contract.workspace = true
 alloy-eips.workspace = true
 alloy-primitives.workspace = true
+alloy-rlp.workspace = true
 alloy-rpc-types-engine.workspace = true
 alloy-sol-types.workspace = true
 base64.workspace = true
