+++ crates/tempo-zone/src/lib.rs
@@ -16,7 +16,6 @@ pub mod evm;
 mod executor;
 pub mod l1;
 pub mod l1_state;
-mod metrics;
 mod node;
 pub mod nonce_keys;
 pub mod payload;
